Use uniform line length in Document
The markdowown document lacked a consistent line lenght of 79 making it
hard to read. Using consistent line length of 79 is a common default.

Markdown links are ugly at line breaks so they were skipped for most
parts. Having the full link on screen is neglectible as long as you have
something to click on.