Substitute Device dictionary with dataclasses
The device dictionary does not respect duplicate protocol entries for one
device, overriding the device dictionary with the latest entry for the same
protocol. This happens if escl devices are present. With the new approach, the
SANE output is mapped to a dataclass, also preserving errorness device by
adding an active=False property to the device.

This adds DataClasses for Device, DeviceOption, DeviceOptionsGroup as well as
Enums for the SANE Types and SANE Units.

Checkboxes are added for Boolean Fields using hidden checkboxes and JavaScript.
The Full Name is now displayed for the Scanners.  Errorness Scanners are greyed
out (disabled) and shown to the user.  Jinja Templates get Full ScanDevice
Class Input.  Also Replace User-Defaulted Custom Printer to the default "None"
Printer that was already present in the code.